backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1
Flask-CORS==4.0.1
PyMySQL==1.1.1
SQLAlchemy==2.0.36
python-dotenv==1.0.1
requests==2.32.3